pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = "Smart Factory Camera Server"
authors = ["Hiram Muñoz <hirammunozr@outlook.com>"]

[tool.poetry.dependencies]
python = ">=3.9, <3.11"
opencv-python = "^4.6.0"
Flask = "^2.1.2"
numpy = "^1.22.4"
onvif-zeep = "^0.2.12"
python-dotenv = "^0.20.0"
matplotlib = "^3.5.2"
Pillow = "^9.1.1"
PyYAML = "^6.0"
scipy = "^1.8.1"
requests = "^2.28.0"
tqdm = "^4.64.0"
pandas = "^1.4.2"
protobuf = "^3.20.0"
toml = "^0.10.2"
seaborn = ">=0.11.0"
ipython = "^8.4.0"
psutil = "^5.9.1"
thop = "^0.1.0"
ipywidgets = "^7.7.0"
torch = "^1.12.0"
torchvision = "^0.13.0"

[tool.poetry.dev-dependencies]
yappi = "^1.3.5"

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu113"
default = false
secondary = false

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
